+++ lib/src/row_cursor.dart
@@ -1,6 +1,6 @@
 import 'package:easy_table/src/row_data.dart';
 import 'package:flutter/material.dart';
 
-/// Signature for a function that defines a row color.
+/// Signature for a function that defines a row cursor.
 /// The theme value will be used if it returns [NULL].
-typedef EasyTableRowColor<ROW> = Color? Function(RowData<ROW> data);
+typedef EasyTableRowCursor<DATA> = MouseCursor? Function(RowData<DATA> data);