+++ lib/src/theme/row_theme_data.dart
@@ -12,7 +12,10 @@ class RowThemeData {
       this.hoverForeground,
       this.dividerThickness = RowThemeDataDefaults.dividerThickness,
       this.dividerColor = RowThemeDataDefaults.dividerColor,
-      this.fillHeight = RowThemeDataDefaults.fillHeight});
+      this.fillHeight = RowThemeDataDefaults.fillHeight,
+      this.cursor = RowThemeDataDefaults.cursor,
+      this.cursorOnTapGesturesOnly =
+          RowThemeDataDefaults.cursorOnTapGesturesOnly});
 
   /// The bottom row color.
   ///
@@ -42,6 +45,11 @@ class RowThemeData {
   /// the color of the rows.
   final bool fillHeight;
 
+  final MouseCursor cursor;
+
+  /// Allow cursor without tap gestures.
+  final bool cursorOnTapGesturesOnly;
+
   @override
   bool operator ==(Object other) =>
       identical(this, other) ||
@@ -53,7 +61,9 @@ class RowThemeData {
           dividerThickness == other.dividerThickness &&
           dividerColor == other.dividerColor &&
           lastDividerVisible == other.lastDividerVisible &&
-          fillHeight == other.fillHeight;
+          fillHeight == other.fillHeight &&
+          cursor == other.cursor &&
+          cursorOnTapGesturesOnly == other.cursorOnTapGesturesOnly;
 
   @override
   int get hashCode =>
@@ -63,7 +73,9 @@ class RowThemeData {
       dividerThickness.hashCode ^
       dividerColor.hashCode ^
       lastDividerVisible.hashCode ^
-      fillHeight.hashCode;
+      fillHeight.hashCode ^
+      cursor.hashCode ^
+      cursorOnTapGesturesOnly.hashCode;
 
   static ThemeRowColor zebraColor(
       {Color? evenColor = const Color(0xFFF5F5F5),
@@ -80,4 +92,6 @@ class RowThemeDataDefaults {
   static const Color dividerColor = Colors.grey;
   static const double dividerThickness = 1;
   static const bool lastDividerVisible = true;
+  static const MouseCursor cursor = SystemMouseCursors.click;
+  static const bool cursorOnTapGesturesOnly = true;
 }